AGRO_CLIMA_APP/requirements.txt
fastapi
uvicorn
dash
dash-bootstrap-components
pandas
httpx
sqlalchemy
alembic
passlib[bcrypt]
python-dotenv

# Additional libraries
selenium
requests
webdriver-manager
# Tools for testing & formatting
pytest
pytest-cov
pytest-asyncio
pylint
black


python-jose
